requirements.txt
flask
qdrant-client
sentence-transformers
pymupdf
google-generativeai
requests